Sliver app bar in config flags page
